azlib_requirements.txt
cachetools==4.2.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
requests==2.26.0